The guessing game
At the beginning I was wondering why the were doing it this way in rust. I guess it was to show all the features that rust have to offer in this game.  There still are some things I see in the code I need to do more research on to get a better understanding.